requirements.txt
# Core dependencies
httpx==0.25.2
aiofiles==23.2.1
pydantic==2.5.0
PyYAML==6.0.1

# Web scraping (browser automation)
playwright==1.40.0

# Data processing
pandas==2.1.4
numpy==1.24.4

# Database
sqlite3  # Built into Python

# Additional utilities
python-dateutil==2.8.2
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Progress bars and CLI
tqdm==4.66.1
click==8.1.7

# Rate limiting and caching
aiocache==0.12.2